requirements.txt
# 基础数据科学
pandas>=2.0.0
numpy>=1.24.0
pyarrow>=14.0.0

# Spark
pyspark>=3.5.0

# 可视化
streamlit>=1.28.0, <1.32.0
plotly>=5.17.0
pydeck>=0.8.0
matplotlib>=3.7.0
seaborn>=0.12.0

# GCP相关（后续用）
google-cloud-storage>=2.10.0
google-cloud-dataproc>=5.2.0
google-cloud-bigquery>=3.11.0

# 工具类
python-dotenv>=1.0.0
pyyaml>=6.0
tqdm>=4.65.0

# 开发工具
black>=23.0.0
flake8>=6.0.0
jupyter>=1.0.0